refactor: :label: Organize button interfaces
